use STRUCTURE for c2p_030 blitter node data
This will allow us to later on have multiple of such structs.
It prepares us for being able to do two blitter c2p runs, one for the upper
half, one for the lower half.
If everything works out nicely, the c2p cpu and blitter work overlaps
and we might be able to reduce the frame latency by some ms.