overlap Blitter and CPU Chunk-to-Planar work by half a screen
After CPU prepared the upper half of the screen for Blitter work,
kick off the Blitter for that half and resume preparing the lower half.
This way, Blitter and CPU work concurrently on the C2P task.

Ideally, this pulls in the time by when the blitter is completely finished by a
few ms and in theory would allow to present the image slightly earlier.
It will not help in overall fps as the rendering is largely CPU bound,
but it might help presenting new frames slightly earlier, reducing perceived lag
a bit.

If we knew that the blitter can race the beam (i.e. all of its work is done in
less than 20ms), we might even be able to somehow use QBSBlit and time the
blitter passes right behind the video beam. This would save the double-buffered
flipping scheme. Right now it could happen that we just miss vsync and then
have to wait for another frame before the flip can actually happen. The fliptask
is often waiting 5-10ms before it can finally flip to the new frame.
It might also be worthwhile investigating if performing the C2P CPU work
during the vertical retrace gets a speedup due to less chipmem contention.
On top of that we could set the blitter nasty bit once all CPU chipmem access
is done to give the blitter full reign.